Programmable SIP Phone Dialing for any Phone Field
Programmable SIP DIAL Integration Implemented. The SIP Server URLs are
in the Virtual Hosts Table. The user table has the user SIP extensions
and passwords for each user.  Note that in the Virtual Hosts SIP URL
field there are variables that can be used as place holders that will be
replaced with the appropriate data for each. They are: [@SIPEXTEN@]
[@SIPPASS@] (Not Encrypted in URL)   [@PHONENUMBER@]

Tarik was the motivator for this, as I was heppy with the simple "tel:
555-555-5555" but now we have a generic SIP solution that can do tel:
555-555-5555 type links or more complex ones say to talk to asterisk
servers directly. Note - there are hooks for VICI already in this
system, but they need to be repopulated with current VICI API commands
to be useful again.