BugFix - MySQL UserName Fixed
The system was broken by updates to the MySQL shared object/libary/dll
that made it so sending the username did not matter, however all the
other fields like server ip and password were not effected. Well, this
commit has changes that allow compiling JAS with and without ODBC, and
allows for compiling with the MySQL4 or MySQL5 Freepascal units. It is
recommended, that you have a separate user for each JAS/JegasCRM
database, and addtionally remove the GRANT PRIVILEGES privilege from
said user so that security can be maintained.